Refactor mini-mecha-code-validator for additional validations
Additional validation checks have been added to the MiniMechaCodeValidator. Specifically, functions to validate that function parameters are unique and an entry function is present in the program model have been implemented. Also, a new constant ENTRY_FUNCTION_NAME was created in a separate file src/utils/constants.ts, which is used for the entry function check. These changes would help ensure the correctness of the code that users create using our mini MECHA programming language and make it error-free.