Add simulator block class and P5.js illustration library
In the simulator, a new Block class was introduced to represent simulation blocks in a more structured way. This class is located in the file public/simulator/block.js. It features dimensions (width, height) and position coordinates (x, y) plus a show() method to display the block.

Also, the P5.js library was added to enable creative coding visualisations. This widely-used JavaScript library will allow us to create graphic and interactive experiences in the simulator. The library is located in the file public/simulator/libraries/p5.js.

These changes bring more structure to block representation, and enhance our ability to create visually engaging and interactive graphics in our simulations.