Updated mini-mecha-code module and added built-in support
This commit introduces several changes in the 'mini-mecha-code' module related to the inclusion of functionalities and the structuration of the code. A new import (MiniMechaCodeWorkspaceManager) and new export types (MiniMechaCodeSharedServices, MiniMechaCodeSharedModule) were added, which will enhance code maintainability and readability. Furthermore, 'mini-mecha-code-validator' now contains a list of built-in MiniMechaCode functions and checks for redeclaration of these built-ins. Simultaneously, support for these built-in functions was extended in the 'main.ts' file along with the introduction of the DslLibraryFileSystemProvider class, the support for a 'builtin' URI schema, and the provision for loading additional documents. This allows us to load the built-in libraries in our language server and offers users tools already set to use, boosting usability.